feat(api): add comprehensive API client and endpoints for attachments, authentication, channels, direct messages, friends, messages, releases, servers, system features, WebRTC, and WebSocket handling
- Implemented attachment handling with upload and URL resolution.
- Added authentication methods including Google OAuth and user management.
- Created channel management functions for creating, renaming, and deleting channels and categories.
- Developed direct messaging capabilities including message sending, editing, and reactions.
- Introduced friend management features for sending and accepting friend requests.
- Established message handling for listing, searching, and pinning messages.
- Integrated release fetching for the latest application updates.
- Built server management functions for server creation, updates, and member management.
- Added system feature retrieval for feature flags.
- Implemented WebRTC token generation and TURN credentials fetching.
- Created WebSocket connection utility for real-time communication.
