+++ apps/web/src/api/releases.ts
@@ -1,0 +1,7 @@
+import { apiFetch } from './client'
+import type { LatestReleaseResponse } from './contracts'
+
+export const releaseApi = {
+    getLatest: () =>
+        apiFetch<LatestReleaseResponse>('/api/releases/latest'),
+}
