+++ apps/web/src/api/system.ts
@@ -1,0 +1,6 @@
+import { apiFetch } from './client'
+import type { SystemFeatures } from './contracts'
+
+export const systemApi = {
+    getFeatures: () => apiFetch<SystemFeatures>('/api/system/features'),
+}
